+++ app.js
@@ -1,24 +1,15 @@
 import 'dotenv/config';
 import express from 'express';
-import {engine} from 'express-handlebars';
 import path from 'node:path';
 import url from 'node:url';
 import dailyLimit from './middlewares/dailyLimit.js';
 import indexRouter from './routes/index.js';
 import mealIntakeApiRouter from './routes/api/meal-intake.js';
+import statusApiRouter from './routes/api/status.js';
 
 const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
 
 const app = express();
-
-// Handlebars設定
-app.engine('hbs', engine({
-  extname: '.hbs',
-  defaultLayout: 'default',
-  layoutsDir: path.join(__dirname, 'views/layout'),
-}));
-app.set('view engine', 'hbs');
-app.set('views', path.join(__dirname, 'views'));
 
 // ミドルウェア
 app.use(express.json({limit: '100mb'}));
@@ -28,6 +19,7 @@ app.use(express.static(path.join(__dirname, 'public')));
 // ルート
 app.use('/', indexRouter);
 app.use('/api', dailyLimit, mealIntakeApiRouter);
+app.use('/api', statusApiRouter);
 
 // 404ハンドラ
 app.use((req, res) => {