+++ client/tailwind.config.js
@@ -1,7 +1,7 @@
 /** @type {import('tailwindcss').Config} */
 module.exports = {
   content: [
-    '../views/**/*.hbs',
+    '../views/**/*.js',
     './src/**/*.js',
   ],
   theme: {